グループ分けを行うクラスを分離
